Moving the sending part of batching RPC into a more modular BatchUtil
